phishing_detectionn/phishing_detection/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
scikit-learn==1.2.2
numpy==1.24.3
pandas==1.5.3
requests==2.31.0
beautifulsoup4==4.12.2
joblib==1.3.1
python-dotenv==1.0.0
openai==1.14.3
httpx==0.27.0
